requirements.txt
pymupdf
arabic-reshaper
python-bidi

